+++ aizynthfinder/aizynthfinder.py
@@ -2,20 +2,36 @@
 """
 from __future__ import annotations
 import time
+from collections import defaultdict
 from typing import TYPE_CHECKING
 
-from deprecated import deprecated
 from tqdm import tqdm
 
 # This must be imported first to setup logging for rdkit, tensorflow etc
 from aizynthfinder.utils.logging import logger
+from aizynthfinder.utils.loading import load_dynamic_class
 from aizynthfinder.context.config import Configuration
-from aizynthfinder.mcts.mcts import SearchTree
-from aizynthfinder.analysis import TreeAnalysis, RouteCollection
-from aizynthfinder.chem import Molecule
+from aizynthfinder.search.mcts import MctsSearchTree
+from aizynthfinder.reactiontree import ReactionTreeFromExpansion
+from aizynthfinder.analysis import (
+    TreeAnalysis,
+    RouteCollection,
+    RouteSelectionArguments,
+)
+from aizynthfinder.chem import Molecule, TreeMolecule, FixedRetroReaction
+from aizynthfinder.search.andor_trees import AndOrSearchTreeBase
 
 if TYPE_CHECKING:
-    from aizynthfinder.utils.type_utils import StrDict, Optional
+    from aizynthfinder.utils.type_utils import (
+        StrDict,
+        Optional,
+        Union,
+        Callable,
+        List,
+        Tuple,
+        Dict,
+    )
+    from aizynthfinder.chem import RetroReaction
 
 
 class AiZynthFinder:
@@ -55,7 +71,7 @@ class AiZynthFinder:
         self.filter_policy = self.config.filter_policy
         self.stock = self.config.stock
         self.scorers = self.config.scorers
-        self.tree: Optional[SearchTree] = None
+        self.tree: Optional[Union[MctsSearchTree, AndOrSearchTreeBase]] = None
         self._target_mol: Optional[Molecule] = None
         self.search_stats: StrDict = dict()
         self.routes = RouteCollection([])
@@ -83,14 +99,16 @@ class AiZynthFinder:
         self.tree = None
         self._target_mol = mol
 
-    def build_routes(self, min_nodes: int = 5, scorer: str = "state score") -> None:
+    def build_routes(
+        self, selection: RouteSelectionArguments = None, scorer: str = "state score"
+    ) -> None:
         """
         Build reaction routes
 
         This is necessary to call after the tree search has completed in order
         to extract results from the tree search.
 
-        :param min_nodes: the minimum number of top-ranked nodes to consider, defaults to 5
+        :param selection: the selection criteria for the routes
         :param scorer: a reference to the object used to score the nodes
         :raises ValueError: if the search tree not initialized
         """
@@ -98,13 +116,20 @@ class AiZynthFinder:
             raise ValueError("Search tree not initialized")
 
         self.analysis = TreeAnalysis(self.tree, scorer=self.scorers[scorer])
-        self.routes = RouteCollection.from_analysis(self.analysis, min_nodes)
+        self.routes = RouteCollection.from_analysis(self.analysis, selection)
 
     def extract_statistics(self) -> StrDict:
         """Extracts tree statistics as a dictionary"""
         if not self.analysis:
             return {}
-        stats = {"target": self.target_smiles, "search_time": self.search_stats["time"]}
+        stats = {
+            "target": self.target_smiles,
+            "search_time": self.search_stats["time"],
+            "first_solution_time": self.search_stats.get("first_solution_time", 0),
+            "first_solution_iteration": self.search_stats.get(
+                "first_solution_iteration", 0
+            ),
+        }
         stats.update(self.analysis.tree_statistics())
         return stats
 
@@ -122,51 +147,9 @@ class AiZynthFinder:
             self.stock.exclude(self.target_mol)
             self._logger.debug("Excluding the target compound from the stock")
 
-        self._logger.debug("Defining tree root: %s" % self.target_smiles)
-        self.tree = SearchTree(root_smiles=self.target_smiles, config=self.config)
+        self._setup_search_tree()
         self.analysis = None
         self.routes = RouteCollection([])
-
-    @deprecated(version="2.1.0", reason="Not supported anymore")
-    def run_from_json(self, params: StrDict) -> StrDict:
-        """
-        Run a search tree by reading settings from a JSON
-
-        :param params: the parameters of the tree search
-        :return: dictionary with all settings and top scored routes
-        """
-        self.stock.select(params["stocks"])
-        self.expansion_policy.select(params.get("policy", params.get("policies", "")))
-        if "filter" in params:
-            self.filter_policy.select(params["filter"])
-        else:
-            self.filter_policy.deselect()
-        self.config.C = params["C"]
-        self.config.max_transforms = params["max_transforms"]
-        self.config.cutoff_cumulative = params["cutoff_cumulative"]
-        self.config.cutoff_number = params["cutoff_number"]
-        self.target_smiles = params["smiles"]
-        self.config.return_first = params["return_first"]
-        self.config.time_limit = params["time_limit"]
-        self.config.iteration_limit = params["iteration_limit"]
-        self.config.exclude_target_from_stock = params["exclude_target_from_stock"]
-        self.config.filter_cutoff = params["filter_cutoff"]
-        self.config.policy_values = params["policy_values"]
-
-        self.prepare_tree()
-        self.tree_search()
-        self.build_routes()
-        if not params.get("score_trees", False):
-            return {
-                "request": self._get_settings(),
-                "trees": self.routes.dicts,
-            }
-
-        self.routes.compute_scores(*self.scorers.objects())
-        return {
-            "request": self._get_settings(),
-            "trees": self.routes.dict_with_scores(),
-        }
 
     def tree_search(self, show_progress: bool = False) -> float:
         """
@@ -177,9 +160,8 @@ class AiZynthFinder:
         """
         if not self.tree:
             self.prepare_tree()
-        assert (
-            self.tree is not None
-        )  # This is for type checking, prepare_tree is creating it.
+        # This is for type checking, prepare_tree is creating it.
+        assert self.tree is not None
         self.search_stats = {"returned_first": False, "iterations": 0}
 
         time0 = time.time()
@@ -188,22 +170,23 @@ class AiZynthFinder:
         time_past = time.time() - time0
 
         if show_progress:
-            pbar = tqdm(total=self.config.iteration_limit)
+            pbar = tqdm(total=self.config.iteration_limit, leave=False)
 
         while time_past < self.config.time_limit and i <= self.config.iteration_limit:
             if show_progress:
                 pbar.update(1)
             self.search_stats["iterations"] += 1
 
-            leaf = self.tree.select_leaf()
-            leaf.expand()
-            while not leaf.is_terminal():
-                child = leaf.promising_child()
-                if child:
-                    child.expand()
-                    leaf = child
-            self.tree.backpropagate(leaf, leaf.state.score)
-            if self.config.return_first and leaf.state.is_solved:
+            try:
+                is_solved = self.tree.one_iteration()
+            except StopIteration:
+                break
+
+            if is_solved and "first_solution_time" not in self.search_stats:
+                self.search_stats["first_solution_time"] = time.time() - time0
+                self.search_stats["first_solution_iteration"] = i
+
+            if self.config.return_first and is_solved:
                 self._logger.debug("Found first solved route")
                 self.search_stats["returned_first"] = True
                 break
@@ -212,43 +195,98 @@ class AiZynthFinder:
 
         if show_progress:
             pbar.close()
+        time_past = time.time() - time0
         self._logger.debug("Search completed")
         self.search_stats["time"] = time_past
         return time_past
 
-    def _get_settings(self) -> StrDict:
-        """Get the current settings as a dictionary"""
-        # To be backward-compatible
-        if (
-            self.expansion_policy.selection
-            and len(self.expansion_policy.selection) == 1
-        ):
-            policy_value = self.expansion_policy.selection[0]
-            policy_key = "policy"
+    def _setup_search_tree(self):
+        self._logger.debug("Defining tree root: %s" % self.target_smiles)
+        if self.config.search_algorithm.lower() == "mcts":
+            self.tree = MctsSearchTree(
+                root_smiles=self.target_smiles, config=self.config
+            )
         else:
-            policy_value = self.expansion_policy.selection  # type: ignore
-            policy_key = "policies"
+            cls = load_dynamic_class(self.config.search_algorithm)
+            self.tree: AndOrSearchTreeBase = cls(
+                root_smiles=self.target_smiles, config=self.config
+            )
 
-        dict_ = {
-            "stocks": self.stock.selection,
-            policy_key: policy_value,
-            "C": self.config.C,
-            "max_transforms": self.config.max_transforms,
-            "cutoff_cumulative": self.config.cutoff_cumulative,
-            "cutoff_number": self.config.cutoff_number,
-            "smiles": self.target_smiles,
-            "return_first": self.config.return_first,
-            "time_limit": self.config.time_limit,
-            "iteration_limit": self.config.iteration_limit,
-            "exclude_target_from_stock": self.config.exclude_target_from_stock,
-            "filter_cutoff": self.config.filter_cutoff,
-            "policy_values": self.config.policy_values,
-        }
 
-        #print parameters to check correct config file
-        print('Parameters:')
-        print(dict_)
+class AiZynthExpander:
+    """
+    Public API to the AiZynthFinder expansion and filter policies
 
-        if self.filter_policy.selection:
-            dict_["filter"] = self.filter_policy.selection
-        return dict_
+    If instantiated with the path to a yaml file or dictionary of settings
+    the stocks and policy networks are loaded directly.
+    Otherwise, the user is responsible for loading them prior to
+    executing the tree search.
+
+    :ivar config: the configuration of the search
+    :ivar expansion_policy: the expansion policy model
+    :ivar filter_policy: the filter policy model
+
+    :param configfile: the path to yaml file with configuration (has priority over configdict), defaults to None
+    :param configdict: the config as a dictionary source, defaults to None
+    """
+
+    def __init__(self, configfile: str = None, configdict: StrDict = None) -> None:
+        self._logger = logger()
+
+        if configfile:
+            self.config = Configuration.from_file(configfile)
+        elif configdict:
+            self.config = Configuration.from_dict(configdict)
+        else:
+            self.config = Configuration()
+
+        self.expansion_policy = self.config.expansion_policy
+        self.filter_policy = self.config.filter_policy
+
+    def do_expansion(
+        self,
+        smiles: str,
+        return_n: int = 5,
+        filter_func: Callable[[RetroReaction], bool] = None,
+    ) -> List[Tuple[FixedRetroReaction, ...]]:
+        """
+        Do the expansion of the given molecule returning a list of
+        reaction tuples. Each tuple in the list contains reactions
+        producing the same reactants. Hence, nested structure of the
+        return value is way to group reactions.
+
+        If filter policy is setup, the probability of the reactions are
+        added as metadata to the reaction.
+
+        The additional filter functions makes it possible to do customized
+        filtering. The callable should take as only argument a `RetroReaction`
+        object and return True if the reaction can be kept or False if it should
+        be removed.
+
+        :param smiles: the SMILES string of the target molecule
+        :param return_n: the length of the return list
+        :param filter_func: an additional filter function
+        :return: the grouped reactions
+        """
+
+        mol = TreeMolecule(parent=None, smiles=smiles)
+        actions, _ = self.expansion_policy.get_actions([mol])
+        results: Dict[Tuple[str, ...], List[FixedRetroReaction]] = defaultdict(list)
+        for action in actions:
+            reactants = action.reactants
+            if not reactants:
+                continue
+            if filter_func and not filter_func(action):
+                continue
+            for name in self.filter_policy.selection or []:
+                if hasattr(self.filter_policy[name], "feasibility"):
+                    _, feasibility_prob = self.filter_policy[name].feasibility(action)
+                    action.metadata["feasibility"] = float(feasibility_prob)
+                    break
+            action.metadata["expansion_rank"] = len(results) + 1
+            unique_key = tuple(sorted(mol.inchi_key for mol in reactants[0]))
+            if unique_key not in results and len(results) >= return_n:
+                continue
+            rxn = next(ReactionTreeFromExpansion(action).tree.reactions())  # type: ignore
+            results[unique_key].append(rxn)
+        return [tuple(reactions) for reactions in results.values()]